+++ CHANGES.rst
@@ -34,6 +34,7 @@ fixes:
 - chore: update Dockerfile python version (#1744)
 - chore: update errbot-backend-slackv3 version to 0.3.2 (#1743)
 - fix: allow webtest to work on python 3.13 (#1729)
+- fix: add command in logged when blocked by cmdfilter (#1631)
 
 
 v6.2.0 (2024-01-01)